requirements.txt
colorama
matplotlib
python-dotenv
requests
SQLAlchemy